Revert to making it a submit
With type=button you can no longer press enter
in a form input to submit the form. With type=submit
it becomes possible again. The form submission is cancelled
because it's not going back to the server, but a submit
event is still triggered, which we listen to in order to trigger a
search.