Simplify the UI
The default combined search is a bit overwhelming and users
often perform simple searches like only keywords, or links.

So we're re-introducing tabs, and moving filters in a 'details'
panel, in order to make the form simpler and to clarify the types
or searches that users may actually need. The 'Mixed' search
type shows the old UI that allows combined searches.