Order results by pagerank
Our cypher queries order pages by pagerank, and the documentation
claims that if a page's pagerank is null, it'll go at the bottom
of the results list. As of the latest neo4j it doesn't seem to be
the case, and null pageranks are listed at the top.

To remedy that, we sort the results ourselves after receiving them.