Use typescript on the server
We're going to migrate and add server code, so we want to enable typescript
Express doesn't do ts very well yet, in particular we can only use commonJS
for linking modules. That's ok except for got which has already dropped support.
Therefore we're switching to axios.

After searching a lot, this has provided a break-through:
https://www.geeksforgeeks.org/how-to-use-express-in-typescript/