Send a 404 when a meta query returns no results
A meta-search with an exact name should always return a result, since the name
is returned from the previous govgraph search.  However we never know what might
happen (in particular the Cypher query might fail) so we need to handle
errors. In this case, when the query returns no result the API sends back a 404
instead of a 500 so that the UI doesn't freeze but instead just doesn't show the
meta box.