Send a custom GA event when the search has occurred
Since this is a single-page app, GA was not notified about
when a user did a search. The only searches GA knew about were
when a user when to a URL that included a search specified in its
query string. However any search triggered by clicking on the search
button weren't send to GA. This commit fixes that by explicitely
sending a custom event every time a search has occurred.