Simplify CI/CD
1. Check in generated JS and CSS files

In theory this is redundant since main.css and main.js are generated from
checked-in source files. But it's simpler and safer not to regenerate them in
CI/CD (because we then no longer need webpack and sass), but to let devs
generate them locally, which they would normally do to test their changes.

2. Move around dev dependencies and don't install them in production