Lower the limit of number of results
The app was crashing for the query `"tribunal decision"` in Cloud Run,
but not when being run locally.  We tried increasing the memory of the
Cloud Run virtual machines from 512MB to 32GB, which didn't help. Then
we tried lowering the limit of the number of results to 1000, which did
work.  Users have previously requested more than 1000 results, so we
then tested a limit of 10,000, which also worked.

There were a couple of different error messages.  One was that a JSON
document was incomplete, and the other was that `Service Unavailable`
isn't valid JSON.

The whole thing is pretty strange, because the queries worked when
running the app on a laptop, querying BigQuery in production.

We also noticed that the results weren't being ordered by descending
number of page views in the query, only in the front end.  That means
that when there were more results than the 50k limit, an arbitrary set
of pages was sent to the app, and then the app sorted them by descending
number of page views.  That bug would be even worse with a lower limit
to the number of pages returned, so this commit introduces `ORDER BY
page_views DESC` in the BigQuery query, so that only relevant pages are
sent to the front end.