Remove test temporarily
The reason why this test no longer passed is because BigQuery now returns
abbreviations. So when looking for "Office for National Statistics" the meta
results will include the correct Organisation but also many abbreviations as
found in the HTML. Since there's always going to be more than one result, the infobox will never show.

With an upcoming version of the frontend, we'll be displaying all the meta
results (in tabs), so we'll re-add this test.