Update deploy script to handle environments
The command to deploy the app is getting complicated, therefore
confusing.  It now tags the cloud run revision with the commit hash, and
the name of the app is easily confused with the name of an environment.

```
gcloud run deploy govuk-knowledge-graph-search \
  --source . \
  --region "europe-west2" \
  --tag commit-`git rev-parse --short HEAD` \
  --project "${PROJECT_ID}"
```

This commit updates the script to prompt the user to choose an
environment, to which it deploys the app.