Log Signon userId and GovSearch sessionId in redis
The /reauth endpoint receives a Signon userId, from which we need to
find related sessions to destroy.  But Redis doesn't support filtering
for keys by value.  So this commit introduces Redis sets of sessionIds
keyed by the Signon userId.

When the /reauth endpoint is used, it finds the set of of sessionIds for
the given userId, pops each sessionId, and destroys the session.