Sanitise SQL of occurrences
- Use SQL parameters instead of interpolated strings, which are insecure
- No need to escape special characters, which is tricky to do without
  affecting the search results
- Support case sensitivity
- Handle the case of zero keywords (which aren't yet supported in the
  front end, but are supported in other parts of the code)
- Return the keywords alongside their counts in the query result, so
  that splitKeywords() doesn't have to be reimplemented in the frontend.

Only the case-sensitive version is tested.  The tests ought to be
overhauled anyway.  They test the logic that constructs SQL
queries, and in order to create test cases, they implement ... logic that
constructs SQL queries.  It would be better to somehow execute the
queries against a static test dataset, which would require a new Google
Cloud project with a BigQuery dataset that has the same name as these
queries refer to.