feat: allow empty keyword search given filters
If we were to allow empty keyword search without any filters, then the
homepage would immediately run a search, returning 10k results from the
entire content of GOV.UK.  So we require at least one filter.

For non-keyword searches, we don't require any filters, but we do
still require a thing (link, organisation, etc.) to search for.

This commit removes the redundant checks in the searchButtonClicked
function, in favour of the ones in the searchState function.  It adds
checks of documentType, publishingStatus and publishingApplication.