feat: update query description on change of sort
The query description includes how the results are sorted, by what
columns, whether ascending or descending, and in what priority.
Previously it was hardcoded to the order in which the API returns the
records. Since the user can now change the displayed sort order, the
query description must be discarded (worse for accessibility) or made
dynamic.

We chose to do this by breaking the "Elm architecture" (described in the
readme), so that instead of handling the table's sortChanged event
with handleEvent(), like other events, we directly update a placeholder
<span>.  This avoids an infinite loop, caused by handleEvent()
eventually triggering another sortChanged event. The alternative seemed
to require a lot of refactoring, which isn't justified by this small
enhancement.

Details:

* Refactor the building of the sort description string into its own
  function, which receives the sort state directly from the event
  listener.
* Don't include the sort order in the query description until the
  results are displayed, and only when they are sorted.
* Add a formatter for "desc" and "asc"