Add publisher_updated_at date column
This is the date that a page was last updated in the Publisher app
(mostly mainstream content).

It is much more meaningful for mainstream content than the
public_updated_at field in the Publishing API and the Content API, which
is rarely used by publishers of mainstream content, and than the
updated_at field of those APIs, which is polluted by editions that are
created for techy reasons rather than editing reasons.

Depends on https://github.com/alphagov/govuk-knowledge-graph-gcp/pull/600